requirements.txt
docx
pytesseract
Pillow
pdf2docx
